Add metadata only to a top doc (assumed by tests)
Consult @treora if it was intended to add CSP & momento
data only on the top document. Implementation used to
generate data URLs for all resources so it was not that
useful there, however if resources are saved in separate files
it might make sense to add metadata everywhere.
At the moment we just do it on the top document to match
assumbtions in tests.