Add Link.resource attribute for now
This should not stay this way. Ideally extract-links could be extracted
into a separate module, not depending on freeze-dry.

We could change the code to either use a generic type for the resource
attribute ('any'?); and possibly rename 'resource' to e.g. 'note' (like
in parse-tools).
Or (probably cleaner) we can change the code to just not add attributes
on the link, but keep them separate (e.g. in a Map<Link,Resource>, or by
wrapping the link in an intermediate object).