Ditch iterators, start recursive processLinks approach
Combining iterators with recursion was never going to work, at least not
without ruining the elegance it was supposed to provide. Instead, use a
callback-based approach, where the callback is invoked for each link to
a subresource and is expected to recurse to handle sub-subresources etc.

The approach is similar to (and inspired by) gozala’s approach in PR #44;
however, it differs in several ways:
- It does not create ‘stub’ resources, because the Link is the item that
  is passed to the callback function.
- Fetching, recursing and updating the link target are explicitly
  performed by the callback function; which may thus require a few more
  lines, but less magic is happening. Also the Resource classes need not
  know anything about these activities, becoming a more passive object.
- Capturing the DOM state is still a separate step doing its own
  recursion, in order to make that happen synchronously. Might change in
  subsequent edits.