WIP Grab frames during subresource crawl, not before
Previously, captureDom would recurse into frames in order to get all
their contents synchronously. Crawling subresources would happen after
this, which works asynchronously, and the reasoning was that the DOM
might change during this asynchronous crawling.

Turns out this was overcautious. The recursive crawling for subresources
can happen partly sync, partly async. If we ensure that the grabbing of
frame contents as well as the recursion into those frames’ frames happen
before the first ‘await’ expression, all frames’ contents will be
obtained synchronously.

Confirmed by MDN[1]: “Top-level code, up to and including the first
await expression (if there is one), is run synchronously.”

This change makes the code simpler: only one recursive walk happens, and
the tactic of first trying to grab it synchronously and falling back to
asynchronous fetching if needed is now cleanly kept in a single place.

This also allows customising the grabbing of frames in the same callback
function.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/async_function