Move frame cloning into DomCloneResource
Puts more logic and opinion into the Resource class, but makes things
somewhat simpler. The ideal would be that the cloned DOM itself contains
the cloned frames in it, but alas the contentDocument attribute of a
frame in a cloned Document appears to be null and is read-only. So we
still keep the cloned subdocument in our custom-made structure of Links.

Now, grabbing the iframe contents recursively is again done in the first,
hard-coded step, rather than being controlled by the processLink callback.
It would be nice to change this again to give control over this recursion.

Also, remove the possibility for custom getDocInFrame implementations to
be async; keeps things simpler for now, could be reintroduced later if
needed (e.g. for browser extensions that need to communicate between
content scripts to work around security restrictions).