Make glob optional in internal functions (‘sub-commit’)
As various 'internal' functions should now become exposed and usable
for customised use cases, having to pass the config everywhere is an
undesirable hassle. Having a default defined in one place was nice and
simple, but now we better add a few lines everywhere to make the more
common case easier.

We could drop support for passing a custom globalThis altogether; but we
still use it in the jest/jsdom tests, and perhaps there are other use
cases for it.

Some usages are unchanged in this commit, if the code is refactored
anyway in the next commit. (this commit is just added to keep that
commit’s diff manageable)